refactor(prometheusreceiver): move targetallocator to internal package
