+++ receiver/prometheusreceiver/internal/targetallocator/manager_test.go
