+++ receiver/prometheusreceiver/internal/targetallocator/testdata/config.yaml
